+++ src/lib/components/features/category/category-detail.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+	import type { Month } from '$lib/utils/month';
+
 	import { getBudget } from '$lib/remote-functions/budget.remote';
 	import { getCategoryById } from '$lib/remote-functions/category.remote';
 
@@ -7,7 +9,11 @@
 	import CategoryEdit from './category-edit.svelte';
 	import CategoryStats from './category-stats.svelte';
 
-	let { categoryId, onDeleted }: { categoryId: string; onDeleted: () => void } = $props();
+	let {
+		categoryId,
+		month,
+		onDeleted
+	}: { categoryId: string; month: Month; onDeleted: () => void } = $props();
 
 	const category = $derived(await getCategoryById({ categoryId }));
 	const budget = $derived(await getBudget(category.budgetId));
@@ -17,7 +23,7 @@
 <div class="grid gap-6 @3xl:grid-cols-2">
 	<CategoryEdit {category} {currency} />
 
-	<CategoryStats {category} {currency} />
+	<CategoryStats {category} {currency} {month} />
 
 	<CategoryArchive {category} {currency} />